Màj Ruben 24/05
Jeu entièrement jouable. Quelques notes :

Les ennemis ont été remaniés. Nouvelle hiérarchie, nouvelles textures :
il ne font plus de moonwalk mais changent d'image en changeant de
direction.

Correction bug affichage des tirs.

Nouveau tutoriel plus agréable et complet.

Fichier des scores contenant les 5 meilleurs scores.

De nouvelles options sont disponibles dans les boites de dialogue de
début et de fin de jeu. En particulier l'affichage des meilleurs scores
et la demande de son nom au joueur ayant effectué un record.

On peut désormais prévisualiser l'endroit où sera placée la défense,
ainsi que sa portee.

Le bug de l'argent négatif semble avoir été corrigé.

Les barrieres ne peuvent plus obstruer totalement le chemin.

Les différentes défenses ne peuvent plus être placées n'importe où.
Leurs infos-bulles ont été mises à jour.

Les récompenses des ennemis, la vie du joueur, son argent, le prix des
défenses et leur ordre ont été ajusté pour pouvoir rendre le jeu jouable
sur une dizaine de niveaux.

SEUL problème restant : il arrive qu'un chemin puisse ne pas être assez
large pour laisser passer les ennemis. De même, il arrive qu'il essaie
de se générer en dehors de la fenêtre. (~1 fois sur 20 au total ...)

+++ Ennemis.java
@@ -17,8 +17,8 @@ public abstract class Ennemis extends Element{
 	public int IDEnnemi;
 	protected int vie;
 	protected int vitesse;
-	public double dposx;
-	public double dposy;
+	public int dposx;
+	public int dposy;
 	protected Arc2D.Double cercle;
 	protected Rectangle limEcran;
 	private final EventListenerList listeners = new EventListenerList();
